# backend/internal/api/api.gen.go
//go:build go1.22

// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package api

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/oapi-codegen/runtime"
)

// Defines values for ErrorCode.
const (
	BADREQUEST      ErrorCode = "BAD_REQUEST"
	INTERNALERROR   ErrorCode = "INTERNAL_ERROR"
	NOTFOUND        ErrorCode = "NOT_FOUND"
	UNAUTHORIZED    ErrorCode = "UNAUTHORIZED"
	VALIDATIONERROR ErrorCode = "VALIDATION_ERROR"
)

// Valid indicates whether the value is a known member of the ErrorCode enum.
func (e ErrorCode) Valid() bool {
	switch e {
	case BADREQUEST:
		return true
	case INTERNALERROR:
		return true
	case NOTFOUND:
		return true
	case UNAUTHORIZED:
		return true
	case VALIDATIONERROR:
		return true
	default:
		return false
	}
}

// Bill defines model for Bill.
type Bill struct {
	Description string `json:"description"`
	Id          int64  `json:"id"`

	// TotalCents Total amount of the bill, in cents.
	TotalCents int64 `json:"totalCents"`
}

// BillSharesResponse defines model for BillSharesResponse.
type BillSharesResponse struct {
	Bill   Bill    `json:"bill"`
	Shares []Share `json:"shares"`

	// TotalBasisPoints Sum of all shares' percentageBasisPoints.
	TotalBasisPoints int32 `json:"totalBasisPoints"`
}

// ErrorCode defines model for ErrorCode.
type ErrorCode string

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Code      ErrorCode     `json:"code"`
	Details   *[]FieldError `json:"details,omitempty"`
	Message   string        `json:"message"`
	RequestId string        `json:"requestId"`
}

// FieldError defines model for FieldError.
type FieldError struct {
	// Field Dotted/indexed path to the offending field, e.g. "shares[2].percentageBasisPoints".
	Field   string `json:"field"`
	Message string `json:"message"`
}

// ReplaceSharesRequest defines model for ReplaceSharesRequest.
type ReplaceSharesRequest struct {
	Shares []ShareInput `json:"shares"`
}

// Share defines model for Share.
type Share struct {
	// AmountCents This share's amount of the bill, in cents.
	AmountCents int64  `json:"amountCents"`
	Id          int64  `json:"id"`
	Name        string `json:"name"`

	// PercentageBasisPoints Percentage in basis points (10000 = 100.00%).
	PercentageBasisPoints int32 `json:"percentageBasisPoints"`
}

// ShareInput defines model for ShareInput.
type ShareInput struct {
	Name                  string `json:"name"`
	PercentageBasisPoints int32  `json:"percentageBasisPoints"`
}

// BillId defines model for BillId.
type BillId = int64

// BadRequest defines model for BadRequest.
type BadRequest = ErrorResponse

// InternalError defines model for InternalError.
type InternalError = ErrorResponse

// NotFound defines model for NotFound.
type NotFound = ErrorResponse

// ValidationError defines model for ValidationError.
type ValidationError = ErrorResponse

// ReplaceBillSharesJSONRequestBody defines body for ReplaceBillShares for application/json ContentType.
type ReplaceBillSharesJSONRequestBody = ReplaceSharesRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// GetBillShares Get the people and their percentages for a bill.
	// (GET /bills/{billId}/shares)
	GetBillShares(w http.ResponseWriter, r *http.Request, billId BillId)
	// ReplaceBillShares Replace the full set of shares for a bill.
	// (PUT /bills/{billId}/shares)
	ReplaceBillShares(w http.ResponseWriter, r *http.Request, billId BillId)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetBillShares operation middleware
func (siw *ServerInterfaceWrapper) GetBillShares(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "billId" -------------
	var billId BillId

	err = runtime.BindStyledParameterWithOptions("simple", "billId", r.PathValue("billId"), &billId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "int64", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "billId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetBillShares(w, r, billId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ReplaceBillShares operation middleware
func (siw *ServerInterfaceWrapper) ReplaceBillShares(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "billId" -------------
	var billId BillId

	err = runtime.BindStyledParameterWithOptions("simple", "billId", r.PathValue("billId"), &billId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "int64", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "billId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ReplaceBillShares(w, r, billId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/bills/{billId}/shares", wrapper.GetBillShares)
	m.HandleFunc(http.MethodPut+" "+options.BaseURL+"/bills/{billId}/shares", wrapper.ReplaceBillShares)

	return m
}

type BadRequestJSONResponse ErrorResponse

type InternalErrorJSONResponse ErrorResponse

type NotFoundJSONResponse ErrorResponse

type ValidationErrorJSONResponse ErrorResponse

type GetBillSharesRequestObject struct {
	BillId BillId `json:"billId"`
}

type GetBillSharesResponseObject interface {
	VisitGetBillSharesResponse(w http.ResponseWriter) error
}

type GetBillShares200JSONResponse BillSharesResponse

func (response GetBillShares200JSONResponse) VisitGetBillSharesResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetBillShares400JSONResponse struct{ BadRequestJSONResponse }

func (response GetBillShares400JSONResponse) VisitGetBillSharesResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type GetBillShares404JSONResponse struct{ NotFoundJSONResponse }

func (response GetBillShares404JSONResponse) VisitGetBillSharesResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)
	_, err := buf.WriteTo(w)
	return err
}

type GetBillShares422JSONResponse struct{ ValidationErrorJSONResponse }

func (response GetBillShares422JSONResponse) VisitGetBillSharesResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(422)
	_, err := buf.WriteTo(w)
	return err
}

type GetBillShares500JSONResponse struct{ InternalErrorJSONResponse }

func (response GetBillShares500JSONResponse) VisitGetBillSharesResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type ReplaceBillSharesRequestObject struct {
	BillId BillId `json:"billId"`
	Body   *ReplaceBillSharesJSONRequestBody
}

type ReplaceBillSharesResponseObject interface {
	VisitReplaceBillSharesResponse(w http.ResponseWriter) error
}

type ReplaceBillShares200JSONResponse BillSharesResponse

func (response ReplaceBillShares200JSONResponse) VisitReplaceBillSharesResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type ReplaceBillShares400JSONResponse struct{ BadRequestJSONResponse }

func (response ReplaceBillShares400JSONResponse) VisitReplaceBillSharesResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type ReplaceBillShares404JSONResponse struct{ NotFoundJSONResponse }

func (response ReplaceBillShares404JSONResponse) VisitReplaceBillSharesResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)
	_, err := buf.WriteTo(w)
	return err
}

type ReplaceBillShares422JSONResponse struct{ ValidationErrorJSONResponse }

func (response ReplaceBillShares422JSONResponse) VisitReplaceBillSharesResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(422)
	_, err := buf.WriteTo(w)
	return err
}

type ReplaceBillShares500JSONResponse struct{ InternalErrorJSONResponse }

func (response ReplaceBillShares500JSONResponse) VisitReplaceBillSharesResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// GetBillShares Get the people and their percentages for a bill.
	// (GET /bills/{billId}/shares)
	GetBillShares(ctx context.Context, request GetBillSharesRequestObject) (GetBillSharesResponseObject, error)
	// ReplaceBillShares Replace the full set of shares for a bill.
	// (PUT /bills/{billId}/shares)
	ReplaceBillShares(ctx context.Context, request ReplaceBillSharesRequestObject) (ReplaceBillSharesResponseObject, error)
}

type StrictHandlerFunc func(ctx context.Context, w http.ResponseWriter, r *http.Request, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	if options.RequestErrorHandlerFunc == nil {
		options.RequestErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	if options.ResponseErrorHandlerFunc == nil {
		options.ResponseErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		}
	}
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// GetBillShares operation middleware
func (sh *strictHandler) GetBillShares(w http.ResponseWriter, r *http.Request, billId BillId) {
	var request GetBillSharesRequestObject

	request.BillId = billId

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetBillShares(ctx, request.(GetBillSharesRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetBillShares")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetBillSharesResponseObject); ok {
		if err := validResponse.VisitGetBillSharesResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// ReplaceBillShares operation middleware
func (sh *strictHandler) ReplaceBillShares(w http.ResponseWriter, r *http.Request, billId BillId) {
	var request ReplaceBillSharesRequestObject

	request.BillId = billId

	var body ReplaceBillSharesJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.ReplaceBillShares(ctx, request.(ReplaceBillSharesRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ReplaceBillShares")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(ReplaceBillSharesResponseObject); ok {
		if err := validResponse.VisitReplaceBillSharesResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}