# contract/openapi.yaml
openapi: 3.0.3
info:
  title: Split Bill API
  description: Splits one bill between people by percentage.
  version: "1.0.0"

paths:
  /bills/{billId}/shares:
    parameters:
      - $ref: "#/components/parameters/BillId"
    get:
      operationId: getBillShares
      summary: Get the people and their percentages for a bill.
      responses:
        "200":
          description: The bill's shares.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillSharesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      operationId: replaceBillShares
      summary: Replace the full set of shares for a bill.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplaceSharesRequest"
      responses:
        "200":
          description: The bill's shares, as saved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillSharesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"

components:
  parameters:
    BillId:
      name: billId
      in: path
      required: true
      description: Positive integer id of the bill.
      schema:
        type: integer
        format: int64
        minimum: 1

  responses:
    BadRequest:
      description: The request was malformed (invalid JSON, invalid path parameter, unknown fields).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The bill does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: The request failed validation (e.g. percentages don't sum to 100.00).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: An unexpected error occurred.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    Bill:
      type: object
      required: [id, description, totalCents]
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        totalCents:
          type: integer
          format: int64
          description: Total amount of the bill, in cents.

    Share:
      type: object
      required: [id, name, percentageBasisPoints, amountCents]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        percentageBasisPoints:
          type: integer
          format: int32
          description: Percentage in basis points (10000 = 100.00%).
        amountCents:
          type: integer
          format: int64
          description: This share's amount of the bill, in cents.

    BillSharesResponse:
      type: object
      required: [bill, shares, totalBasisPoints]
      properties:
        bill:
          $ref: "#/components/schemas/Bill"
        shares:
          type: array
          items:
            $ref: "#/components/schemas/Share"
        totalBasisPoints:
          type: integer
          format: int32
          description: Sum of all shares' percentageBasisPoints.

    ShareInput:
      type: object
      required: [name, percentageBasisPoints]
      properties:
        name:
          type: string
        percentageBasisPoints:
          type: integer
          format: int32

    ReplaceSharesRequest:
      type: object
      required: [shares]
      properties:
        shares:
          type: array
          items:
            $ref: "#/components/schemas/ShareInput"

    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - BAD_REQUEST
        - NOT_FOUND
        - UNAUTHORIZED
        - INTERNAL_ERROR

    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          description: 'Dotted/indexed path to the offending field, e.g. "shares[2].percentageBasisPoints".'
        message:
          type: string

    ErrorResponse:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        requestId:
          type: string